requirements.txt
Django==3.2.14
Pillow==8.3.2 
django-environ

dj-database-url
dj-static
python-decouple
gunicorn
psycopg2-binary

djangorestframework
django-filter 

beautifulsoup4
selenium
chromedriver
chromedriver-binary==121.0.6167.16.0